feat: fold replay order is receipt order (rowid) — O(n²) ingest dissolves

Merged as an internal waypoint, NOT a released contract.

decision:design/declared-fold-order amendment #4 (01M070N08HX69QAEQ3TRPY72CQ)
supersedes the two-mode design this branch was written against. The canonical
arrival substrate rewrite lands on top of this, not beside it.

Why merge it anyway: this branch introduces no per-store receipt-mode flag. It
unconditionally makes fold order the order this store received facts and demotes
(ts,id) to an explicit read lens — which IS the arrival model's semantics,
implemented one layer too low. Its authority for "order received" is the sqlite
rowid, which today equals JSONL line order. Holding it out of tree would make the
rewrite re-fight every (ts,id)-as-fold-order pin still on main and rebuild
replay_from plus checkpoint dispatch from zero with no base-arm benchmark.

Kept: Spec.replay_from, the VertexHandle checkpoint machinery, the same-runner
benchmark arms, the lens conformance area, Rule 17's ratchet.
Discarded as doctrine: R1 (merge.py:83, "the ceremony DEFINES the merged store's
receipt order") — under arrival, merge never touches custody.

No tag, no release. The one persisted-store contract break is spent on the model
actually believed. Git history is allowed to contain the stepping stone; the
persisted data model is not.

+++ libs/engine/src/engine/ceremony.py
@@ -194,7 +194,7 @@ class DeclarationUpdatePreview:
     documents: tuple  # proposed projection, Document.as_json dicts
     proposed_fingerprint: str
     old_store_fingerprint: str | None  # store projection at plan; None pre-genesis
-    expected_head: tuple[float, str] | None  # CAS token; None in genesis mode
+    expected_head: tuple[int, str] | None  # CAS token; None in genesis mode
     authority: str  # "file" | "store"
     applicable: bool
     reason: str
@@ -301,7 +301,7 @@ def plan_declaration_update(
         *,
         mode: str,
         changes: tuple = (),
-        expected_head: tuple[float, str] | None = None,
+        expected_head: tuple[int, str] | None = None,
         old_store_fingerprint: str | None = None,
         authority: str,
         applicable: bool,
